requirements.txt
asgiref==3.8.1
attrs==25.3.0
certifi==2025.8.3
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.4.0
cryptography==43.0.3
defusedxml==0.8.0rc2
distlib==0.3.9
Django==5.0.11
django-cors-headers==4.7.0
django-filter==24.3
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
drf-yasg==1.21.15
filelock==3.18.0
gunicorn==25.1.0
h11==0.16.0
html-testRunner==1.2.1
idna==3.10
inflection==0.5.1
Jinja2==3.1.6
MarkupSafe==3.0.2
msgpack==1.1.0
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==26.0
pillow==12.1.1
platformdirs==4.3.7
psycopg2-binary==2.9.11
pycparser==2.22
PyJWT==2.9.0
PySocks==1.7.1
python-decouple==3.8
python-dotenv==1.2.2
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.3
redis==5.2.0
requests==2.32.3
requests-oauthlib==2.0.0
selenium==4.35.0
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sortedcontainers==2.4.0
sqlparse==0.5.1
stripe==14.4.1
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
tzdata==2024.2
Unipath==1.1
uritemplate==4.2.0
urllib3==2.5.0
virtualenv==20.30.0
virtualenvwrapper-win==1.2.7
websocket-client==1.8.0
whitenoise==6.11.0
wsproto==1.2.0
